fix: code improvements
