+++ .eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  'extends': 'eslint:recommended',
+  rules: {    
+    'no-empty': 'error'
+  },
+}
