+++ .github/CONTRIBUTING.md
@@ -1,0 +1,7 @@
+# How to contribute
+
+1. Fork the project
+2. Create your feature branch (`git checkout -b feaure/my-amazing-feature`)
+3. Commit your changes (`git commit -m 'feature: added amazing feature'`)
+4. Push to the branch (`git push origin feature/my-amazing-feature`)
+5. Open a Pull Request
