+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vgs/collect-js",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "VGS Collect.js script loading module",
   "license": "MIT",
   "main": "dist/index.js",
@@ -16,7 +16,7 @@
     "start": "tsdx watch",
     "build": "tsdx build",
     "test": "tsdx test",
-    "lint": "tsdx lint --fix",
+    "lint": "tsdx lint",
     "prepare": "tsdx build"
   },
   "peerDependencies": {},
@@ -37,11 +37,11 @@
     "husky": "^4.3.0",
     "tsdx": "^0.13.3",
     "tslib": "^2.0.1",
-    "typescript": "^4.0.2"
+    "typescript": "^4.0.2",
+    "uuid": "^8.3.0"
   },
   "dependencies": {
     "@types/uuid": "^8.3.0",
-    "axios": "^0.20.0",
-    "uuid": "^8.3.0"
+    "axios": "^0.20.0"
   }
 }
